+++ db/migrate/20191130141938_add_geocoder_errors_to_location.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddGeocoderErrorsToLocation < ActiveRecord::Migration[6.0]
+  def change
+    add_column :locations, :geocoder_errors, :string, array: true, default: []
+  end
+end