+++ spec/controllers/location_queries_controller_spec.rb
@@ -3,24 +3,23 @@
 require 'rails_helper'
 
 RSpec.describe LocationQueriesController, type: :controller do
-  include Dry::Effects::Handler.Resolve
-
   around do |spec|
-    DependencyResolverMiddleware.new(nil).testing do
+    DependencyResolver.testing do
       spec.run
     end
   end
 
   describe 'GET /location' do
-    it 'returns a list of locations' do
+    before do
       get :show
+    end
 
+    it 'returns ok' do
       expect(response).to have_http_status(:ok)
-      body = response.body
-      expect(body).not_to be_empty
-      parsed = JSON.parse(body)
-      expect(parsed).to have_key('features')
-      expect(parsed['features'].size).to eq(6)
+    end
+
+    it 'returns a list of locations' do
+      expect(response).to include_json(features: ->(f) { f.size == 6 })
     end
   end
 
@@ -39,7 +38,8 @@ RSpec.describe LocationQueriesController, type: :controller do
         { q: '' } => :missing_query_param,
         {} => :missing_query_param,
         { q: '}' } => :invalid_json,
-        { q: '{"type":"Point","coordinates":[8.3]}' } => :invalid_json
+        { q: '{"type":"Point","coordinates":[8.3]}' } => :invalid_json,
+        { q: '{"type":"Polygon","coordinates":[[[12.32,48.10],[12.06,48.06],[12.54,46.66],[12.32,48.10]]]}' } => :invalid_json
       }
 
       params_to_error.each do |params, error|
@@ -49,6 +49,37 @@ RSpec.describe LocationQueriesController, type: :controller do
           expect(response).to have_http_status(:bad_request)
           expect(response.body).to eq(JSON.dump(error: error))
         end
+      end
+    end
+  end
+
+  describe 'GET /location/by-id' do
+    before do
+      get :by_id, params: { id: id }
+    end
+
+    context 'when ID is not found' do
+      let(:id) { SecureRandom.uuid }
+
+      it 'returns 404' do
+        expect(response).to have_http_status(:not_found)
+      end
+    end
+
+    context 'when ID belongs to a Location' do
+      let(:name) { 'Austin' }
+      let(:id) { Location.create!(name: name).id }
+
+      it 'returns 200' do
+        expect(response).to have_http_status(:ok)
+      end
+
+      it 'returns name' do
+        expect(response).to include_json(name: name)
+      end
+
+      it 'returns only whitelisted params' do
+        expect(response).to be_json_response_with_keys(%w[name inside lonlat geocoder_errors])
       end
     end
   end